Fix bug in addEntry
I fixed a bug in the form processing for adding an entry. Turns out I forgot to add the data from the file to the $contacts array at the beginning of the file. That was fixed and as an added bonus, the form now outputs prettier code in JSON format.